Refactor(jira): Move classes into new packages.
