+++ alert-issuetracker/src/main/java/com/synopsys/integration/alert/issuetracker/jira/server/JiraServerCreateIssueTestAction.java
@@ -28,9 +28,9 @@ import org.slf4j.LoggerFactory;
 import com.google.gson.Gson;
 import com.synopsys.integration.alert.issuetracker.IssueCreator;
 import com.synopsys.integration.alert.issuetracker.IssueCreatorTestAction;
-import com.synopsys.integration.alert.issuetracker.IssueTrackerContext;
 import com.synopsys.integration.alert.issuetracker.IssueTrackerService;
 import com.synopsys.integration.alert.issuetracker.TransitionValidator;
+import com.synopsys.integration.alert.issuetracker.config.IssueTrackerContext;
 import com.synopsys.integration.alert.issuetracker.jira.cloud.JiraCloudCreateIssueTestAction;
 import com.synopsys.integration.alert.issuetracker.jira.common.util.JiraTransitionHandler;
 import com.synopsys.integration.alert.issuetracker.jira.server.util.JiraServerTransitionHandler;
