+++ alert-issuetracker/src/main/java/com/synopsys/integration/alert/issuetracker/jira/server/JiraServerService.java
@@ -28,8 +28,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.gson.Gson;
-import com.synopsys.integration.alert.issuetracker.IssueTrackerContext;
 import com.synopsys.integration.alert.issuetracker.IssueTrackerService;
+import com.synopsys.integration.alert.issuetracker.config.IssueTrackerContext;
 import com.synopsys.integration.alert.issuetracker.exception.IssueTrackerException;
 import com.synopsys.integration.alert.issuetracker.jira.common.JiraConstants;
 import com.synopsys.integration.alert.issuetracker.jira.server.util.JiraServerIssueHandler;
@@ -57,7 +57,7 @@ public class JiraServerService extends IssueTrackerService {
 
     @Override
     public IssueTrackerResponse sendMessage(IssueTrackerContext context, Collection<IssueTrackerRequest> requests) throws IntegrationException {
-        JiraServerProperties jiraProperties = (JiraServerProperties) context.getIssueTrackerConfig();
+        JiraServerProperties jiraProperties = context.getIssueTrackerConfig();
         JiraServerServiceFactory jiraServerServiceFactory = jiraProperties.createJiraServicesServerFactory(logger, getGson());
         PluginManagerService jiraAppService = jiraServerServiceFactory.createPluginManagerService();
         logger.debug("Verifying the required application is installed on the Jira server...");
