+++ alert-issuetracker/src/main/java/com/synopsys/integration/alert/issuetracker/message/IssueCommentRequest.java
@@ -1,6 +1,5 @@
 package com.synopsys.integration.alert.issuetracker.message;
 
-import com.synopsys.integration.alert.issuetracker.IssueContentModel;
 import com.synopsys.integration.alert.issuetracker.IssueProperties;
 import com.synopsys.integration.alert.issuetracker.OperationType;
